+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Prefix the change with one of these keywords:
 - Fixed: BackDrop component is now decoupled from React Portals
 - Fixed: Button in the SearchBar component is now aligned with the TextField
 - Fixed: Toggle now hides elements inside properly
+- Deprecated: `onDarkBackground` prop in `Link`. Use `darkBackground` instead.
 
 ## [0.20.0]